requirements.txt
pyautogen
yfinance